版本检测更新，系统每天自动检测一次，关于我信页面

+++ RKWXT/LoginAndRegist/Model/GainModel.m
@@ -23,7 +23,7 @@
         }else{
             NSInteger smsID = [[dic objectForKey:@"sms_id"] integerValue];
             WXTUserOBJ *userDefault = [WXTUserOBJ sharedUserOBJ];
-            [userDefault setSmsID:smsID];
+            [userDefault setSmsID:(int)smsID];
             
             if (_delegate && [_delegate respondsToSelector:@selector(gainNumSucceed)]){
                 [_delegate gainNumSucceed];
